De-duplicate `Solution` data structure
